user auth

+++ laragig/app/Http/Controllers/ListingController.php
@@ -46,6 +46,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         
@@ -60,6 +62,11 @@ class ListingController extends Controller
     }
 
     public function update(Request $request, Listing $listing){
+
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id())
+            abort('403', 'Unauthorized action');
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -83,7 +90,16 @@ class ListingController extends Controller
 
     //Delete listing
     public function destroy(Listing $listing){
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id())
+            abort('403', 'Unauthorized action');
+        
         $listing->delete();
         return redirect('/')->with('message', 'Listing Deleted Successfully');
+    }
+
+    //Manage Listing
+    public function manage(){
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
